solve Anagrams algorithm problem
Alphabetic Anagrams

We can then assign a number to every word, based on where it falls in an alphabetically sorted list of all words made up of the same group of letters. One way to do this would be to generate the entire list of words and find the desired one, but this would be slow if the word is long.

in https://www.codewars.com/kata/alphabetic-anagrams/train/java